🔧 clean up Research Data page
